requirements.txt
streamlit
Pillow
pandas
requests
pypdf
yt-dlp